collected all schemas in dedicated schema.ts file and removed validators.ts. Imported schemas to orders/actions.ts
